refactor: extract getEnhancedContentForSession to shared utility
