+++ apps/desktop/src/contexts/search/engine/indexing.ts
@@ -6,7 +6,12 @@ import {
   createSessionSearchableContent,
 } from "./content";
 import type { Index } from "./types";
-import { collectCells, toNumber, toTrimmedString } from "./utils";
+import {
+  collectCells,
+  getEnhancedContentForSession,
+  toNumber,
+  toTrimmedString,
+} from "./utils";
 
 export function indexSessions(db: Index, store: PersistedStore): void {
   const fields = [
@@ -33,23 +38,6 @@ export function indexSessions(db: Index, store: PersistedStore): void {
       created_at: toNumber(row.created_at),
     });
   });
-}
-
-function getEnhancedContentForSession(
-  store: PersistedStore,
-  sessionId: string,
-): string {
-  const contents: string[] = [];
-  store.forEachRow("enhanced_notes", (rowId: string, _forEachCell) => {
-    const noteSessionId = store.getCell("enhanced_notes", rowId, "session_id");
-    if (noteSessionId === sessionId) {
-      const content = store.getCell("enhanced_notes", rowId, "content");
-      if (typeof content === "string" && content) {
-        contents.push(content);
-      }
-    }
-  });
-  return contents.join(" ");
 }
 
 export function indexHumans(db: Index, store: PersistedStore): void {
